+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `reactionrole` command (messiest code I've ever seen but if I come up with a better solution I will tidy it all up)
+
 ## [1.4.5-alpha] - 2021-11-17
 ### Added
 - `ping` command (messages customizable in `config.js`)